requirements.txt.txt
streamlit
pandas
googletrans==4.0.0-rc1
biopython
stripe
openpyxl
